python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emacs-agent"
version = "0.1.0"
description = "A Python-based AI assistant for Emacs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "google-adk",
    "playwright",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["emacs_agent"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
